requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
aiogram==3.3.0
aiofiles==23.2.1
Pillow==10.1.0 